rename wizzard to wizard
