+++ client/version_control/ui/workspace_wizard/css/style.css
